requirements.txt
streamlit
openai
pandas
python-dotenv
gTTS
openpyxl